Api: use float for vehicle emission

+++ app/api/v01/entities/vehicle.rb
@@ -24,7 +24,7 @@ class V01::Entities::VehicleWithoutVehicleUsage < Grape::Entity
   expose(:contact_email, documentation: { type: String })
   expose(:ref, documentation: { type: String })
   expose(:name, documentation: { type: String })
-  expose(:emission, documentation: { type: Integer })
+  expose(:emission, documentation: { type: Float })
   expose(:consumption, documentation: { type: Integer })
   expose(:capacity, documentation: { type: Integer, desc: 'Deprecated, use capacities instead.' }) { |m|
     if m.capacities && m.customer.deliverable_units.size == 1
